Python script to monitor changes on a website
This script monitors a website for any change and alerts the user via email. The email alert can be modified, so can the time interval between accessing the website. The hash of the entire website is computed and compared in the next iteration of the script, if the hash values do not match, the website's state has been altered and alert is issued. This was made to run in the background or to be run on a cloud server, depending on the case. A simple script to automate our endless refreshing the uni's result page.